% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{addhealth}
\alias{addhealth}
\title{A portion of the AddHealth dataset pulled from an archived UCI site}
\format{
A list of \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} graph objects, each corresponding
to one of the original 84 communities.

Each community graph is a directed graph with node table:
\itemize{
\item \code{sex} (factor): Sex of student, coded as \code{male} or \code{female}. Possibly missing.
\item \code{race} (factor): Race of student, code as \code{white}, \code{black}, \code{hispanic}, \code{asian}, or \code{mixed/other}. Possibly missing.
\item \code{grade} (integer): Grade of student, between 6 and 12. Possibly missing.
\item \code{school} (factor): For communities with sister schools, the first school is coded as factor \code{A} and the second as \code{B}.
}

and edge table:
\itemize{
\item \code{from} (int): Id of student declaring friend
\item \code{to} (int): Id of student declared as friend
\item \code{weight} (int): A number from 1 to 6. See details for an explanation.
}
}
\source{
Originally at \url{http://moreno.ss.uci.edu/data.html}, but
ultimately accessed via \url{https://web.archive.org/web/20210115130726/http://moreno.ss.uci.edu/data.html}
}
\usage{
addhealth
}
\description{
Per the original documentation:
}
\details{
The ADD HEALTH data are constructed from the in-school questionnaire;
90,118 students representing 84 communities took this survey in 1994-95.
Some communities had only one school; others had two. Where there are two
schools in a community students from one school were allowed to name
friends in the other, the "sister school."

Each student was given a paper-and-pencil questionnaire and a copy of a
roster listing every student in the school and, if the community had two
schools, the student s provided with the roster of the "sister" school.
The name generator asked about five male and five female friends separately.
The question was, "List your closest (male/female) friends. List your best
(male/female) friend first, then your next best friend, and so on.
(girls/boys) may include (boys/girls) who are friends and (boy/girl)
friends."

For each friend named, the student was asked to check off whether he/she
participated in any of five activities with the friend. These activities
were:
\enumerate{
\item you went to (his/her) house in the last seven days.
\item you met (him/her) after school to hang out or go somewhere in the last seven days.
\item you spent time with (him/her) last weekend.
\item you talked with (him/her) about a problem in the last seven days.
\item you talked with (him/her) on the telephone in the last seven days.
}

These activities were summed to create a valued network. Ties range in
value from 1, meaning the student nominated the friend but reported no
activities, to 6, meaning the student nominated the friend and reported
participating in all five activities with the friend.

Communities 1 and 48 are omitted. Community 1 is omitted due to data that
does not follow the original coding scheme. Community 48 is omitted because
the node attribute data was not available online at the time we pulled it.
}
\references{
Moody, James, "Peer influence groups: identifying dense clusters in large networks," Social Networks, 2001, 23: 261-283.
}
\keyword{datasets}
