% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{ase_specials}
\alias{ase_specials}
\alias{U,}
\alias{V,}
\alias{US,}
\alias{VS}
\alias{U}
\alias{US}
\alias{V}
\title{Helper functions for including node embeddings in formulas}
\usage{
U(A, rank, ...)

US(A, rank, ...)

V(A, rank, ...)

VS(A, rank, ...)
}
\arguments{
\item{A}{A \code{\link[=matrix]{matrix()}} or \code{\link[Matrix:Matrix]{Matrix::Matrix()}} object.}

\item{rank}{Rank of desired decomposition.}

\item{...}{
  Arguments passed on to \code{\link[irlba:irlba]{irlba::irlba}}
  \describe{
    \item{\code{nv}}{number of right singular vectors to estimate.}
    \item{\code{nu}}{number of left singular vectors to estimate (defaults to \code{nv}).}
    \item{\code{maxit}}{maximum number of iterations.}
    \item{\code{work}}{working subspace dimension, larger values can speed convergence at the cost of more memory use.}
    \item{\code{reorth}}{if \code{TRUE}, apply full reorthogonalization to both SVD bases, otherwise
only apply reorthogonalization to the right SVD basis vectors; the latter case is cheaper per
iteration but, overall, may require more iterations for convergence. Automatically \code{TRUE}
when \code{fastpath=TRUE} (see below).}
    \item{\code{tol}}{convergence is determined when \eqn{\|A^TU - VS\| < tol\|A\|}{||A^T U - VS|| < tol*||A||},
and when the maximum relative change in estimated singular values from one iteration to the
next is less than \code{svtol = tol} (see \code{svtol} below),
where the spectral norm ||A|| is approximated by the
largest estimated singular value, and U, V, S are the matrices corresponding
to the estimated left and right singular vectors, and diagonal matrix of
estimated singular values, respectively.}
    \item{\code{v}}{optional starting vector or output from a previous run of \code{irlba} used
to restart the algorithm from where it left off (see the notes).}
    \item{\code{right_only}}{logical value indicating return only the right singular vectors
(\code{TRUE}) or both sets of vectors (\code{FALSE}). The right_only option can be
cheaper to compute and use much less memory when \code{nrow(A) >> ncol(A)} but note
that obtained solutions typically lose accuracy due to lack of re-orthogonalization in the
algorithm and that \code{right_only = TRUE} sets \code{fastpath = FALSE} (only use this option
for really large problems that run out of memory and when \code{nrow(A) >> ncol(A)}).
Consider increasing the \code{work} option to improve accuracy with \code{right_only=TRUE}.}
    \item{\code{verbose}}{logical value that when \code{TRUE} prints status messages during the computation.}
    \item{\code{scale}}{optional column scaling vector whose values divide each column of \code{A};
must be as long as the number of columns of \code{A} (see notes).}
    \item{\code{center}}{optional column centering vector whose values are subtracted from each
column of \code{A}; must be as long as the number of columns of \code{A} and may
not be used together with the deflation options below (see notes).}
    \item{\code{shift}}{optional shift value (square matrices only, see notes).}
    \item{\code{mult}}{DEPRECATED optional custom matrix multiplication function (default is \code{\%*\%}, see notes).}
    \item{\code{fastpath}}{try a fast C algorithm implementation if possible; set \code{fastpath=FALSE} to use the
reference R implementation. See the notes for more details.}
    \item{\code{svtol}}{additional stopping tolerance on maximum allowed absolute relative change across each
estimated singular value between iterations.
The default value of this parameter is to set it to \code{tol}. You can set \code{svtol=Inf} to
effectively disable this stopping criterion. Setting \code{svtol=Inf} allows the method to
terminate on the first Lanczos iteration if it finds an invariant subspace, but with less certainty
that the converged subspace is the desired one. (It may, for instance, miss some of the largest
singular values in difficult problems.)}
    \item{\code{smallest}}{set \code{smallest=TRUE} to estimate the smallest singular values and associated
singular vectors. WARNING: this option is somewhat experimental, and may produce poor
estimates for ill-conditioned matrices.}
  }}
}
\description{
Quickly compute top \code{rank} singular vectors of a matrix \code{A} and
return them in a matrix. Especially useful in formulas in regression
objects.
}
\details{
In all cases, computes a top-\code{rank} decomposition of \code{A}, such that
\eqn{A \approx U S V^T}.
\itemize{
\item \code{U} returns \eqn{U}
\item \code{US} returns \eqn{U S^{1/2}}
\item \code{V} returns \eqn{V}
\item \code{VS} returns \eqn{V S^{1/2}}
}
}
\seealso{
nodelm, vsp_specials
}
