% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-loss.R
\name{coef.mediator}
\alias{coef.mediator}
\title{Title}
\usage{
\method{coef}{mediator}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{other arguments.}
}
\value{
TODO
}
\description{
Title
}
\examples{

mu <- model_mediator_uninformative(n = 100, k = 5, dim_c = 3)

coef(mu)


}
