% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{glasgow}
\alias{glasgow}
\title{Teenage Friends and Lifestyle Data}
\format{
A list of three \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} directed graphs objects,
representing the same network measured at three different points in
time.

TODO: describe the tidygraphs
}
\source{
Retrieved from
\url{https://www.stats.ox.ac.uk/~snijders/siena/Glasgow_data.zip} on
June 8, 2022.
}
\usage{
glasgow
}
\description{
The canonical "Glasgow" data, as reported in the \code{siena} R package
for network data analysis. \code{latentnetmediate} simply re-exports this data
in a more convenient format. See the \code{siena} data document for
gory details.
}
\references{
\url{https://www.stats.ox.ac.uk/~snijders/siena/Glasgow_data.htm}
}
\keyword{datasets}
