% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Leukemia cases by census tract in New York}
\format{
An undirected \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} graph object, made up
of a node table and an edge table.

The node table has columns:
\itemize{
\item \code{population} (numeric): TODO
\item \code{cases} (numeric): TODO
\item \code{censustract_fips} (list of character vectors): TODO
}

and the edge table has columns:
\itemize{
\item \code{from} (int): Id of source census tract
\item \code{to} (int): Id of neighboring census tract
}
}
\source{
Extracted from the \code{SpatialEpi} package on July 29, 2022. See
\link[SpatialEpi:NYleukemia]{SpatialEpi::NYleukemia} for details, as well as
Wakefield and Kim (2013).
}
\usage{
data(leukemia)
}
\description{
Leukemia cases by census tract in New York
}
\references{
Wakefield, J., and A. Kim. "A Bayesian Model for Cluster
Detection." Biostatistics 14, no. 4 (September 1, 2013):
752–65. https://doi.org/10.1093/biostatistics/kxt001.
}
\keyword{datasets}
