% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-loss.R
\name{mediator_coef}
\alias{mediator_coef}
\alias{coef.uninformative}
\alias{coef.block}
\alias{coef.informative}
\alias{coef.perfect}
\title{Title}
\usage{
\method{coef}{uninformative}(object, ...)

\method{coef}{block}(object, ...)

\method{coef}{informative}(object, ...)

\method{coef}{perfect}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{other arguments.}
}
\value{
TODO
}
\description{
Title
}
\examples{

mu <- model_mediator_uninformative(n = 100, k = 5, dim_c = 3)

coef(mu)


}
