% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-canonical-interventions.R
\name{model_canonical}
\alias{model_canonical}
\title{Title}
\usage{
model_canonical(n, k, ...)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{...}{
  Arguments passed on to \code{\link[=model_mediator_informative]{model_mediator_informative}}
  \describe{
    \item{\code{ztheta_0}}{TODO}
    \item{\code{ztheta_t}}{TODO}
    \item{\code{ztheta_c}}{TODO}
    \item{\code{ztheta_tc}}{TODO}
  }}
}
\value{
TODO
}
\description{
Title
}
\examples{

set.seed(26)

mrdpg <- model_canonical(n = 100, k = 5)

graph <- sample_tidygraph(mrdpg)
graph

m_fit <- nodelm(US(A, 5) ~ . - name - y - 1, graph = graph)
o_fit <- nodelm(y ~ . - name - 1 + US(A, 5), graph = graph)

m_fit
o_fit

}
