% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-contagion-basic.R
\name{model_contagion_basic}
\alias{model_contagion_basic}
\title{Title}
\usage{
model_contagion_basic(n, k = 5, dim_c = 3)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{dim_w}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

set.seed(26)

b <- model_contagion_basic(n = 100, k = 5, dim_c = 3)

graph <- sample_tidygraph(b)
graph


}
