% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-causal.R, R/model-mediator.R
\name{model_informative_controls}
\alias{model_informative_controls}
\title{Title}
\usage{
model_informative_controls(
  n,
  k = 5,
  prob_trt = 0.5,
  theta_0 = NULL,
  theta_t = NULL,
  theta_c = NULL,
  theta_tc = NULL,
  beta_0 = NULL,
  beta_t = NULL,
  beta_c = NULL,
  beta_x = NULL
)

model_informative_controls(
  n,
  k = 5,
  prob_trt = 0.5,
  theta_0 = NULL,
  theta_t = NULL,
  theta_c = NULL,
  theta_tc = NULL,
  beta_0 = NULL,
  beta_t = NULL,
  beta_c = NULL,
  beta_x = NULL
)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{prob_trt}{TODO}

\item{theta_0}{TODO}

\item{theta_t}{TODO}

\item{theta_c}{TODO}

\item{theta_tc}{TODO}

\item{beta_0}{TODO}

\item{beta_t}{TODO}

\item{beta_c}{TODO}

\item{beta_x}{TODO}
}
\value{
TODO

TODO
}
\description{
Title

Title
}
