% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-interference-basic.R
\name{model_interference_basic}
\alias{model_interference_basic}
\title{Title}
\usage{
model_interference_basic(n, k = 5, dim_w = 3)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{dim_w}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

set.seed(26)

b <- model_interference_basic(n = 100, k = 5, dim_w = 3)

graph <- sample_tidygraph(b)
graph


}
