% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{model_mediator_block}
\alias{model_mediator_block}
\title{Title}
\usage{
model_mediator_block(n, k = 5, dim_c = 3)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{dim_c}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

library(estimatr)
library(ggplot2)
library(tidyr)

set.seed(26)

mblock <- model_mediator_block(n = 100, k = 5, dim_c = 3)

plot_expected_a_pre_trt(mblock)
plot_expected_a_post_trt(mblock)
plot_expected_a_pre_post_diff(mblock)

graph <- sample_tidygraph(mblock)
graph

fit <- nodelm(US(A, 5) ~ trt1 + trt2 + trt3 + trt4 + trt5 + C1 + C2 + C3 , graph = graph)

true_coefs <- tibble(
  outcome = 1:5,
  `(Intercept)` = drop(mblock$theta_0),
  trt1 = mblock$theta_t[1, ],
  trt2 = mblock$theta_t[2, ],
  trt3 = mblock$theta_t[3, ],
  trt4 = mblock$theta_t[4, ],
  trt5 = mblock$theta_t[5, ],
  C1 = mblock$theta_c[1, ],
  C2 = mblock$theta_c[2, ],
  C3 = mblock$theta_c[3, ]
) |>
  tidyr::pivot_longer(
    cols = -outcome,
    names_to = "term",
    values_to = "estimate"
  )

tidy(fit) |>
  ggplot(aes(x = term, ymin = conf.low, y = estimate, ymax = conf.high)) +
  geom_pointrange() +
  geom_point(
    data = true_coefs,
    aes(x = term, y = estimate),
    color = "red",
    shape = 4,
    size = 2,
    inherit.aes = FALSE
  ) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  coord_flip() +
  facet_wrap(
    vars(outcome),
    labeller = "label_both"
  ) +
  theme_classic()

}
