% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator-block2.R
\name{model_mediator_block2}
\alias{model_mediator_block2}
\title{Title}
\usage{
model_mediator_block2(n, k = 5, expected_degree = NULL)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{expected_degree}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

set.seed(26)

mblock <- model_mediator_block2(n = 100, k = 5)

graph <- sample_tidygraph(mblock)
graph

fit <- nodelm(US(A, 5) ~ . - name - 1, graph = graph)
fit

}
