% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator-informative.R
\name{model_mediator_informative}
\alias{model_mediator_informative}
\title{Title}
\usage{
model_mediator_informative(
  n,
  k = 5,
  ztheta_0 = NULL,
  ztheta_t = NULL,
  ztheta_c = NULL,
  ztheta_tc = NULL,
  expected_degree = NULL
)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{ztheta_0}{TODO}

\item{ztheta_t}{TODO}

\item{ztheta_c}{TODO}

\item{ztheta_tc}{TODO}

\item{expected_degree}{If specified, the desired expected degree
of the graph. Specifying \code{expected_degree} simply rescales \code{S}
to achieve this. Defaults to \code{NULL}. Do not specify both
\code{expected_degree} and \code{expected_density} at the same time.}
}
\value{
TODO
}
\description{
Title
}
\examples{

inf <- model_mediator_informative(n = 100, k = 5)

graph <- sample_tidygraph(inf)
graph

nodelm(US(A, 5) ~ . - name - 1, graph = graph)

coef(inf)

# should be very close to zero
inf$zX - inf$zC_true

inf$zX - inf$zC_obs
inf$W - inf$zC_obs

}
