% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{model_mediator_perfect}
\alias{model_mediator_perfect}
\title{Title}
\usage{
model_mediator_perfect(
  n,
  k = 5,
  ztheta_0 = NULL,
  ztheta_t = NULL,
  ztheta_c = NULL,
  ztheta_tc = NULL,
  expected_degree = 10
)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{ztheta_0}{TODO}

\item{ztheta_t}{TODO}

\item{ztheta_c}{TODO}

\item{ztheta_tc}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

perf <- model_mediator_perfect(n = 100)

graph <- sample_tidygraph(perf)
graph

igraph::as_adj(graph)


coef(perf)

perf$X - perf$C

}
