% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator-uninformative.R
\name{model_mediator_uninformative}
\alias{model_mediator_uninformative}
\title{Title}
\usage{
model_mediator_uninformative(n, k = 5, dim_w = 3)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{dim_w}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

set.seed(26)

mu <- model_mediator_uninformative(n = 100, k = 5, dim_w = 3)

graph <- sample_tidygraph(mu)
graph

coef(mu)

fit <- nodelm(US(A, 5) ~ . - name - 1, graph = graph)
fit

}
