% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-canonical-interventions.R
\name{model_uninformative}
\alias{model_uninformative}
\title{Title}
\usage{
model_uninformative(n, k, ...)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{...}{
  Arguments passed on to \code{\link[=model_mediator_uninformative]{model_mediator_uninformative}}
  \describe{
    \item{\code{dim_w}}{TODO}
  }}
}
\value{
TODO
}
\description{
Title
}
\examples{

set.seed(26)

mrdpg <- model_uninformative(n = 100, k = 5)

graph <- sample_tidygraph(mrdpg)
graph

m_fit <- nodelm(US(A, 5) ~ . - name - y - 1, graph = graph)
o_fit <- nodelm(y ~ . - name - 1 + US(A, 5), graph = graph)

m_fit
o_fit

mrdpg$nde
coef(o_fit)

coef(m_fit)
coef(o_fit)

}
