% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{netmediate}
\alias{netmediate}
\title{Estimate network direct and indirect effects via a spectral method}
\usage{
netmediate(graph, formula, rank)
}
\arguments{
\item{graph}{A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} object.}

\item{formula}{Details about the nodal design matrix. Of the form
outcome ~ nodal_formula. For now, no interactiosn or fancy stuff
are allowed in the formula.}

\item{rank}{Integer describing desired rank of embedding dimension.}
}
\value{
A \code{network_mediation} object.
}
\description{
\strong{TODO}: Allow embedding L_tau(A) and L(A) in addition to A
}
\examples{

library(tidygraph)

smokeint <- smoking |>
  mutate(
    smokes_int = as.integer(smokes) - 1
  )

netmed <- smokeint |>
  netmediate(smokes_int ~ sex + 0, rank = 7)

netmed

}
