% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-plot.R
\name{plot_expected_a_pre_trt}
\alias{plot_expected_a_pre_trt}
\alias{plot_expected_a_post_trt}
\alias{plot_expected_a_pre_post_diff}
\alias{plot_x_pre_post_diff}
\title{Title}
\usage{
plot_expected_a_pre_trt(model)

plot_expected_a_post_trt(model)

plot_expected_a_pre_post_diff(model)

plot_x_pre_post_diff(model)
}
\arguments{
\item{model}{TODO}
}
\value{
TODO
}
\description{
Title
}
