% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-plot.R
\name{plot_intervention_on_x}
\alias{plot_intervention_on_x}
\title{TODO}
\usage{
plot_intervention_on_x(model, post_multiply = FALSE)
}
\arguments{
\item{model}{TODO}

\item{post_multiply}{TODO}
}
\value{
TODO
}
\description{
TODO
}
