% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{reddit}
\alias{reddit}
\title{Top level gender-flaired Reddit comments}
\format{
A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} bipartite graph object, made up
of a node table and an edge table.

The node table has columns:
\itemize{
\item \code{type} (logical): A logical indicator of whether a node corresponds to
a word or a reddit post. Used internally by \code{igraph} -- see \code{node_type}
for an easier to use alternative.
\item \code{name} (character): Unique node identifier. There is one node in the
graph for each reddit post, and for each word used in a reddit post.
Posts are identified by a number. Words are identified by tokenized
words. Tokenized of raw top-level comment text was performed with
\code{tidytext::unnest_tokens(..., token = "tweets")}.
\item \code{node_type} (character): Either \code{"post"} or \code{"word"}.
\item \code{author_gender}: Either \code{"female"} or \code{"male"}, based on user
flairs. Only available for post nodes.
\item \code{score} (integer): The number of upvotes minus the number of downvotes
received by a given post. Only available for post nodes.
\item \code{subreddit} (character): One of \code{"keto"}, \code{"okcupid"} or \code{"childfree"}.
Only available for post nodes.
\item \code{author_pseudonym} (character): An author identifier that is consistent
across posts. Only available for post nodes.
}

and the edge table has columns:
\itemize{
\item \code{from} (int): Id of post
\item \code{to} (int): Id of word
\item \code{weight} (double): Number of times a word was used in a post.
}
}
\source{
Downloaded from \url{https://archive.org/details/reddit_posts_2018} on
June 6, 2022.
}
\usage{
reddit
}
\description{
Reddit is an online forum divided into topic-specific subforums called
'subreddits. We consider three subreddits: \code{keto}, \code{okcupid}, and
\code{childfree}. In these subreddits, we identify users whose username
flair includes a gender label (usually 'M' or 'F'). We collect all top-level
comments from these users in 2018. We use each comment's text and score,
the number of likes minus dislikes from other users.
The dataset includes 90k comments in the selected subreddits.
}
\details{
See \url{https://github.com/blei-lab/causal-text-embeddings} for a replication
package for Veitch et al (2020). See \url{https://github.com/blei-lab/causal-text-embeddings/blob/master/src/reddit/data_cleaning/BigQuery_get_data}
in particular for additional data details.
}
\references{
Veitch, Victor, Dhanya Sridhar, and David M Blei.
"Adapting Text Embeddings for Causal Inference." In Proceedings of the
36 Th Conference on Uncertainty in Artificial Intelligence (UAI),
124:10, 2020.
}
\keyword{datasets}
