% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{sample_tidygraph}
\alias{sample_tidygraph}
\title{Sample from a mediated RDPG object}
\usage{
sample_tidygraph(model, ...)
}
\arguments{
\item{model}{A \code{mediated_rdpg} object}

\item{...}{Ignored.}
}
\value{
A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} object.
}
\description{
Sample from a mediated RDPG object
}
