% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{sensitivity_curve}
\alias{sensitivity_curve}
\title{Estimate mediated effects for a variety of embedding dimensions}
\usage{
sensitivity_curve(graph, formula, max_rank, ranks_to_consider = 10)
}
\arguments{
\item{graph}{A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} object.}

\item{formula}{Details about the nodal design matrix. Of the form
outcome ~ nodal_formula. For now, no interactiosn or fancy stuff
are allowed in the formula.}

\item{max_rank}{Maximum rank to consider (integer).}

\item{ranks_to_consider}{How many distinct ranks to consider (integer).
Optional, defaults to 10.}
}
\value{

}
\description{
Estimate mediated effects for a variety of embedding dimensions
}
\examples{


rank_curve <- smoking |>
  mutate(
    smokes_int = as.integer(smokes) - 1
  ) |>
 sensitivity_curve(smokes_int ~ sex + 0, max_rank = 7)

rank_curve
plot(rank_curve)

}
