% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking in a network of chinese college students}
\format{
A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} graph object, made up of a node table
and an edge table.

The node table has columns:
\itemize{
\item \code{sex} (factor): Sex of student, coded as \code{male} or \code{female}.
\item \code{smokes} (factor): Whether student smokes. Levels are \code{smokes} and \verb{doesn't smoke}.
}

and the edge table has columns:
\itemize{
\item \code{from} (int): Id of student declaring friend
\item \code{to} (int): Id of student declared as friend
}
}
\source{
Obtained from Haiyan Liu over email on April 28, 2022.
}
\usage{
smoking
}
\description{
Data collected by the Lab for Big Data Methodology at the
University of Notre Dame in 2017. The participants were 162 students in a
4-year college in China. There were 90 female and 72 male students.
Their average age was 21.64 years (SD = 0.86). During the data collection,
each student indicated whether the other students were his/her friends
or not. We symmetrized the friendship network using the strongest relation
(i.e., if one of two students indicated the other as a friend,
we assumed they were mutual friends).
}
\references{
Liu, Haiyan, Ick Hoon Jin, Zhiyong Zhang, and Ying Yuan.
"Social Network Mediation Analysis: A Latent Space Approach."
Psychometrika 86, no. 1 (March 2021): 272–98.
https://doi.org/10.1007/s11336-020-09736-z.
}
\keyword{datasets}
