% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{vsp_specials}
\alias{vsp_specials}
\alias{Z,}
\alias{Y}
\alias{Z}
\title{Helper functions for including rotated node embeddings in formulas}
\usage{
Z(A, rank, ..., degree_normalize = FALSE)

Y(A, rank, ..., degree_normalize = FALSE)
}
\arguments{
\item{A}{A \code{\link[=matrix]{matrix()}} or \code{\link[Matrix:Matrix]{Matrix::Matrix()}} object.}

\item{rank}{Rank of desired decomposition.}

\item{...}{
  Arguments passed on to \code{\link[vsp:vsp]{vsp::vsp}}
  \describe{
    \item{\code{x}}{Either a graph adjacency matrix, \link[igraph:aaa-igraph-package]{igraph::igraph} or
\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}. If \code{x} is a \link{matrix} or \link[Matrix:Matrix]{Matrix::Matrix}
then \code{x[i, j]} should correspond to the edge going from node \code{i}
to node \code{j}.}
  }}
}
\description{
In all cases, computes a top-\code{rank} decomposition of \code{A}, such that
\eqn{A \approx U S V^T = Z B Y^T}, where \eqn{Z} is \eqn{U} after
varimax rotation and \eqn{Y} is \eqn{V} after varimax rotation. See
\code{\link[vsp:vsp]{vsp::vsp()}} for details.
}
\seealso{
nodelm, ase_specials
}
