% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{model_mediator_informative}
\alias{model_mediator_informative}
\title{Title}
\usage{
model_mediator_informative(
  n,
  k = 5,
  theta_0 = NULL,
  theta_t = NULL,
  theta_c = NULL,
  theta_tc = NULL
)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{theta_0}{TODO}

\item{theta_t}{TODO}

\item{theta_c}{TODO}

\item{theta_tc}{TODO}
}
\value{
TODO
}
\description{
Title
}
