% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{model_mediator_informative}
\alias{model_mediator_informative}
\title{Title}
\usage{
model_mediator_informative(
  n,
  k = 5,
  ztheta_0 = NULL,
  ztheta_t = NULL,
  ztheta_c = NULL,
  ztheta_tc = NULL
)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{ztheta_0}{TODO}

\item{ztheta_t}{TODO}

\item{ztheta_c}{TODO}

\item{ztheta_tc}{TODO}
}
\value{
TODO
}
\description{
Title
}
