% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{model_mediator_informative}
\alias{model_mediator_informative}
\title{Title}
\usage{
model_mediator_informative(
  n,
  k = 5,
  prob_trt = 0.5,
  theta_0 = NULL,
  theta_t = NULL,
  theta_c = NULL,
  theta_tc = NULL,
  beta_0 = NULL,
  beta_t = NULL,
  beta_c = NULL,
  beta_x = NULL
)
}
\arguments{
\item{beta_0}{TODO}

\item{beta_t}{TODO}

\item{beta_c}{TODO}

\item{beta_x}{TODO}
}
\value{
TODO
}
\description{
Title
}
