% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator-perfect.R
\name{model_mediator_perfect}
\alias{model_mediator_perfect}
\title{Title}
\usage{
model_mediator_perfect(
  n,
  k = 5,
  ztheta_0 = NULL,
  ztheta_t = NULL,
  ztheta_c = NULL,
  ztheta_tc = NULL,
  expected_degree = NULL
)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{ztheta_0}{TODO}

\item{ztheta_t}{TODO}

\item{ztheta_c}{TODO}

\item{ztheta_tc}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

ztheta_tc <- rbind(
  c(-1, 2, 0, 0, 0), # treated in block one become twice as central and jump to block two
  c(0, 0, 0, 0, 0),
  c(0, 1, -1, 0, 0), # treated in block three jump to block two
  c(0, 0, 0, 0, 0),
  c(0, 0, 0, 0, 0)
)

perf <- model_mediator_perfect(n = 100, k = 5, ztheta_tc = ztheta_tc)

}
