% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mediator.R
\name{model_mediator_uninformative}
\alias{model_mediator_uninformative}
\title{Title}
\usage{
model_mediator_uninformative(n, k = 5, dim_c = 3)
}
\arguments{
\item{n}{TODO}

\item{k}{TODO}

\item{dim_c}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

library(estimatr)
library(ggplot2)
library(dplyr)

set.seed(26)

mu <- model_mediator_uninformative(n = 100, k = 5, dim_c = 3)

plot_expected_a_pre_trt(mu)
plot_expected_a_post_trt(mu)
plot_expected_a_pre_post_diff(mu)

graph <- sample_tidygraph(mu)
graph

fit <- nodelm(US(A, 5) ~ trt + C1 + C2 + C3 , graph = graph)

true_coefs <- tibble(
  outcome = 1:5,
  `(Intercept)` = drop(mu$theta_0),
  trt = drop(mu$theta_t),
  C1 = mu$theta_c[1, ],
  C2 = mu$theta_c[2, ],
  C3 = mu$theta_c[3, ]
) |>
  tidyr::pivot_longer(
    cols = -outcome,
    names_to = "term",
    values_to = "estimate"
  )

tidy(fit) |>
  ggplot(aes(x = term, ymin = conf.low, y = estimate, ymax = conf.high)) +
  geom_pointrange() +
  geom_point(
    data = true_coefs,
    aes(x = term, y = estimate),
    color = "red",
    shape = 4,
    size = 2,
    inherit.aes = FALSE
  ) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  coord_flip() +
  facet_wrap(
    vars(outcome),
    labeller = "label_both"
  ) +
  theme_classic()

}
