% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-plot.R
\name{plot_expected_a_pre_trt}
\alias{plot_expected_a_pre_trt}
\alias{plot_expected_a_post_trt}
\alias{plot_expected_a_pre_post_diff}
\alias{plot_z_pre_post_diff}
\alias{plot_sparse_adjacency_matrix}
\alias{plot_tidygraph_matrix}
\alias{plot_tidygraph_matrix_rownormalized}
\title{Title}
\usage{
plot_expected_a_pre_trt(model)

plot_expected_a_post_trt(model)

plot_expected_a_pre_post_diff(model)

plot_z_pre_post_diff(model)

plot_sparse_adjacency_matrix(A)

plot_tidygraph_matrix(graph)

plot_tidygraph_matrix_rownormalized(graph)
}
\arguments{
\item{model}{TODO}

\item{A}{TODO}

\item{graph}{TODO}
}
\value{
TODO
}
\description{
Title
}
